Model Building Page
This code runs as the Model Building Page of the application and helps the user build various Classification and Regression models on the chosen independent and dependent variables of the input training and testing size, tuning the hyperparameters according to the model built. It also allows the user to evaluate the model using numerous metrics.